==> analysis_final/2_loan_outcome_determination.ipynb <==
import os
import pandas as pd 
import glob
import random 
import numpy as np
import matplotlib.pyplot as plt
%matplotlib inline 
import math

pd.set_option('display.max_columns',200)
pd.set_option('display.max_rows',680)

## Read in data: 

os.listdir()

Step 1: Inspect data

trx = pd.read_csv('diaries_trx_trunc_loans_start_bal.csv', low_memory = False, dtype = {'new_account_ids': 'object'})

trx.shape

trx.loc[trx['unique_accnts']==1].trx_type_desc.value_counts()

trx.loc[trx['unique_accnts']==1].trx_type_code.value_counts()

Step 2: Let's look at regularity a bit

# for idx, acct_id in enumerate(list(trx.account_ids.unique())):
#     print(idx, acct_id)
#     loan_df = trx.loc[trx['account_ids']==acct_id]    
#     loan_df = loan_df.copy()
#     loan_df.sort_values("trx_stdtime_days_acc", inplace=True)
    
#     #subset down only to payments
#     payment_idxs = [idx for idx, purpose in enumerate(list(loan_df['trx_prx_purpose'])) if 'Payment' in purpose]
#     payment_df= loan_df.iloc[payment_idxs]
    
#     #figure out a vector of time between payments
#     time_btw_payments = np.diff(np.array(payment_df.trx_stdtime_days_acc))
    
#     print('time btw payments', time_btw_payments)
#     print('std of time btw payments', np.std(time_btw_payments))
#     print('mean of payment amount', np.mean(payment_df.trx_value_kes))
#     print('-------------------------')

Step 3: Test out a single loan: 

acct_id = list(trx.new_account_ids.unique())[0]

loan_df = trx.loc[trx['new_account_ids']==acct_id]    
loan_df = loan_df.copy()
loan_df.sort_values("trx_stdtime_days_acc", inplace=True)

loan_df.trx_stdtime_days_acc.max()

loan_df.shape

loan_df.columns

days_observed = loan_df.trx_stdtime_days_acc.max()- loan_df.trx_stdtime_days_acc.min()
x_days = 150

if days_observed >= x_days: 
    thresh_day = loan_df.trx_stdtime_days_acc.min()+150 
    loan_x_days = loan_df.loc[loan_df['trx_stdtime_days_acc'] <= thresh_day].copy()
else:     
    loan_x_days = loan_df.copy()

len(list(loan_x_days.bal_evol)) > 0

balance_at_x_days = list(loan_x_days.bal_evol)[-1]
balance_at_x_days

balance_at_x_days <= 0

Step 4: Put it into a function: has a loan been paid back in x_days? 

def find_loans_paid_back_in_x_days(df, x_days):
    loan_categories = pd.DataFrame()
    for idx, acct_id in enumerate(list(df.new_account_ids.unique())):
#         print(idx, acct_id)
        
        loan_df = df.loc[df['new_account_ids']==acct_id]    
        loan_df = loan_df.copy()
        loan_df.sort_values("trx_stdtime_days_acc", inplace=True)
        owner_id = list(loan_df.m_ids_owner)[0]
        key = str('paid_back_in_'+ str(x_days)+'_days')
        
        
        ## find the number of days the loan was observed: 
        days_observed = loan_df.trx_stdtime_days_acc.max()- loan_df.trx_stdtime_days_acc.min()
        
        ## if the loan is obeserved for more than 150 days, we need to just look and see if the last balance
        ## on the last transaction day on before (first day+150) is <= 0
        ## then we create a new df that is less than or equal to our thresh day, otherwise we just use that same df 
        
        if days_observed >= x_days: 
            thresh_day = loan_df.trx_stdtime_days_acc.min()+150 
            loan_x_days = loan_df.loc[loan_df['trx_stdtime_days_acc'] <= thresh_day].copy()
        else:     
            loan_x_days = loan_df.copy()
        
        ## if the balance evolution is greater than 0
        if len(list(loan_x_days.bal_evol)) > 0:
        
            balance_at_x_days = list(loan_x_days.bal_evol)[-1]

            if balance_at_x_days <= 0: 
                loan_categories_line = {'new_account_ids': str(acct_id), 'm_ids_owner': owner_id, key: 1}
            else: 
                loan_categories_line = {'new_account_ids': str(acct_id), 'm_ids_owner': owner_id, key: 0}
        else:
            loan_categories_line = {'new_account_ids': str(acct_id), 'm_ids_owner': owner_id, key: 'too_few_borrowings'}
        
        loan_categories = loan_categories.append(loan_categories_line, ignore_index = True)        
    return loan_categories

trx.shape

loans_paid_back_in_150_days_df = find_loans_paid_back_in_x_days(trx, 150)

loans_paid_back_in_150_days_df

Step 5: Create function that takes in full trx dataset, a dataset that says whether they pay back in x days, x_days and a threshold for "regularity" of payments 

def find_regular_loans(df, loans_paid_back_in_x_days_df, thresh, x_days):
    
    # df is df of all trx of loans
    # thresh is the cutoff for num days between loans 
    # x_days = days of observation of the loan 

    loan_categories = pd.DataFrame()
    
    for idx, acct_id in enumerate(list(df.new_account_ids.unique())):
        
#         print(idx, acct_id)
        
        loan_df = df.loc[df['new_account_ids']==acct_id]    
        loan_df = loan_df.copy()
        loan_df.sort_values("trx_stdtime_days_acc", inplace=True)
        owner_id = list(loan_df.m_ids_owner)[0]
        key1 = 'every_payment_under_'+str(thresh)+'_days'
        

        #subset down only to payments
        payment_idxs = [idx for idx, purpose in enumerate(list(loan_df['trx_prx_purpose'])) if 'Payment' in purpose]
        payment_df= loan_df.iloc[payment_idxs]
        
        # calculate the array of time between payments: 
        time_btw_payments = np.diff(np.array(payment_df.trx_stdtime_days_acc))
        
        # starting time 
        starting_time = list(loan_df.trx_stdtime_days_acc)[0]
        ending_time = list(loan_df.trx_stdtime_days_acc)[-1]
        num_to_add_to_time_btw_payments = x_days+starting_time - ending_time
        
        key_lookup = 'paid_back_in_'+str(x_days)+'_days'
        
        paid_back_x_days = loans_paid_back_in_x_days_df.loc[loans_paid_back_in_x_days_df['new_account_ids']==acct_id][key_lookup]
        paid_back_x_days = int(paid_back_x_days)
        if paid_back_x_days == 0: 
            np.append(time_btw_payments, num_to_add_to_time_btw_payments)
        else: 
            time_btw_payments = time_btw_payments
        
#         print('time btw payments', time_btw_payments)
#         print('std of time btw payments', np.std(time_btw_payments))
#         print('mean of payment amount', np.mean(payment_df.trx_value_kes))
#         print('-------------------------')

        ints_over_thresh = [x for x in  time_btw_payments if x > thresh]
        
        if len(ints_over_thresh)>0: 
            loan_categories_line = {'new_account_ids': str(acct_id), 'm_ids_owner': owner_id, key1: 0}
        else: 
            loan_categories_line = {'new_account_ids': str(acct_id), 'm_ids_owner': owner_id, key1: 1}
        loan_categories = loan_categories.append(loan_categories_line, ignore_index = True)        
    return loan_categories

regular_loanpayers = find_regular_loans(trx, loans_paid_back_in_150_days_df, 33, 150)

Step 6: Test the merge 

merged_outcomes = loans_paid_back_in_150_days_df.merge(regular_loanpayers, 'outer', ['new_account_ids', 'm_ids_owner'])

merged_outcomes.shape

def create_outcomes(trx_df, thresh, x_days): 
    
    x = find_loans_paid_back_in_x_days(trx_df, x_days)
    y = find_regular_loans(trx_df, x, thresh, x_days)
    
    z = x.merge(y, 'outer', ['new_account_ids', 'm_ids_owner'])
    return z 

loan_outcomes = create_outcomes(trx, 33, 150)

loan_outcomes.head(20)

loan_outcomes.shape

loan_outcomes.to_csv('loan_outcomes_33_150.csv', index = False)
