==> analysis_final/4_merge_predictors_with_outcome.ipynb <==
import pandas as pd 
import os 
import glob 
from collections import Counter

pd.set_option('display.max_columns',200)
pd.set_option('display.max_rows',680)

# Merging

glob.glob('../datasets/*csv')

predictors = pd.read_csv('../datasets/cons_wb_go_edu_demos_inc_pov_rem.csv', dtype = {'m_ids_owner': 'object'})

for col in predictors.columns:
    print(col)

glob.glob('*csv')

outcome = pd.read_csv('individual_outcomes_33_150.csv', dtype = {'m_ids_owner': 'object'})

print(predictors.shape)
print(outcome.shape)

predictors.head()

outcome.head()

 outcome = outcome.drop(['loans'], axis = 1)

outcome.shape

data = outcome.merge(predictors, on = 'm_ids_owner', how = 'left')

data.shape

data.to_csv('all_data.csv', index = False)
