==> analysis_final/5_format_data.ipynb <==
import pandas as pd 
import os 
import glob 
from collections import Counter

# Merging

glob.glob("*csv")

pd.set_option('display.max_rows', 500)
pd.set_option('display.max_columns', 500)
pd.set_option('display.width', 1000)

data = pd.read_csv('all_data.csv', dtype = {'m_ids_owner': 'object'})

data.columns

# Formatting: 

data.shape

## Dealing with the categorical variables: 

data.dtypes[data.dtypes == 'object']

#### Dropping age group b/c we have age itself: 

data = data.drop(['dem_i_age_5yrgrp'], axis = 1)

#### Reformatting the edu_i_attain entries: 

data.edu_i_attain.value_counts()

new_edu = []
for idx, row in data.iterrows(): 
    if 'Primary' in str(row['edu_i_attain']): 
        new_edu.append(2)
    elif 'Secondary' in str(row['edu_i_attain']): 
        new_edu.append(3)
    elif 'Post-secondary' in str(row['edu_i_attain']): 
        new_edu.append(4)
    elif 'education' in str(row['edu_i_attain']): 
        new_edu.append(0)
    else: 
        new_edu.append(1)

Counter(new_edu)

data.edu_i_attain = new_edu

data.edu_i_attain.value_counts()

#### Reformatting the dem_i_male entries: 

data.dem_i_male.value_counts()

new_male = []
for idx, row in data.iterrows(): 
    if 'Female' in str(row['dem_i_male']): 
        new_male.append(0)
    else: 
        new_male.append(1)

data.dem_i_male = new_male

data.dtypes[data.dtypes == 'object']

#### Reformatting the dem_i_relhead entries: 

data.dem_i_relhead.value_counts()

new_relhead = []
for idx, row in data.iterrows(): 
    if 'Household' in str(row['dem_i_relhead']): 
        new_relhead.append('hh_head')
    elif 'Husband' in str(row['dem_i_relhead']):
        new_relhead.append('h_or_w')
    elif 'Son' in str(row['dem_i_relhead']):
        new_relhead.append('s_or_d')    
    else: 
        new_relhead.append('parent')
        

data.dem_i_relhead = new_relhead

new_relhead = pd.get_dummies(data.dem_i_relhead, prefix = 'dem_i_relhead')

data = pd.concat([data.reset_index(drop=True), new_relhead], axis=1)

data = data.drop('dem_i_relhead', axis = 1)

# data.columns

data.dtypes[data.dtypes == 'object']

#### Reformatting the dem_i_marstat entries: 

data.dem_i_marstat.value_counts()

new_marstat = []
for idx, row in data.iterrows(): 
    if 'Married' in str(row['dem_i_marstat']): 
        new_marstat.append('married')
    elif 'Widowed' in str(row['dem_i_marstat']):
        new_marstat.append('widowed')
    elif 'Separated' in str(row['dem_i_marstat']):
        new_marstat.append('separated')    
    else: 
        new_marstat.append('never_married')
        

data.dem_i_marstat = new_marstat

new_marstat = pd.get_dummies(data.dem_i_marstat)

data = pd.concat([data.reset_index(drop=True), new_marstat], axis=1)

data = data.drop('dem_i_marstat', axis = 1)

data.dtypes[data.dtypes == 'object']

#### Reformatting the dem_i_tribe entries: 

data.dem_i_tribe.value_counts()

new_tribe = pd.get_dummies(data.dem_i_tribe, prefix = 'tribe')

data = pd.concat([data.reset_index(drop=True), new_tribe], axis=1)

data = data.drop('dem_i_tribe', axis = 1)

data.dtypes[data.dtypes == 'object']

data["hh_inf_lighting_source"].value_counts()

data.dtypes[data.dtypes == 'object']

###  deleting "HH" from m_ids_owner

data.loc[data['m_ids_owner']=="HH"]

data = data.drop(35)

data.to_csv('data_merged_nocatvars.csv', index = False)

## I'm only going to take those variables that are complete from here; later once Dinara's ds is fixed we can do more with imputation and stuff: 

data = pd.read_csv('data_merged_nocatvars.csv')

data.describe().loc['count']

data[data["inc_i_meanpti"].isnull()]

There seems to be one person that has a loan but has no income information: 65134382764500000.

cols_to_remove = data.describe().count().index[data.describe().loc['count']<135]

data = data.drop(cols_to_remove, axis= 1)

data.describe().loc['count'].value_counts()

data.to_csv('data_merged_noMissingData.csv', index = False)

## Removing unnecessary variables: 

This shows how many columns have 139 entries

data = pd.read_csv('data_merged_nocatvars.csv')

data.shape

pd.DataFrame(data.describe()).loc['count'].value_counts()

98 cols have 138 entries (e.g. no NA values) 

If we don't have information for over half the population, going to just delete those columns: 

data.describe().loc['count']<69

cols_to_remove = data.describe().count().index[data.describe().loc['count']<69]

data = data.drop(cols_to_remove, axis= 1)

data.describe().loc['count'].value_counts()

data.shape
