==> analysis_final/6_train_test_split.ipynb <==
import pandas as pd
import glob
import os
from sklearn.model_selection import train_test_split

glob.glob('*.csv')

data = pd.read_csv('data_merged_noMissingData.csv')

data.shape

data.head()

data = data.loc[data['binarize_score'] != 999]

data.binarize_score.value_counts()

data.shape

data.head()

labels = data.binarize_score 
data = data.drop(['binarize_score'], axis = 1)

x_train, x_test, y_train, y_test = train_test_split(data, labels, stratify = labels, random_state=42, test_size = 0.2 )

x_train.head()

y_train.head()

y_train = pd.DataFrame(y_train)
y_train.columns = ['owner_score_bin']

y_test = pd.DataFrame(y_test)
y_test.columns = ['owner_score_bin']

x_train.to_csv('binary_x_train.csv', index =0)
x_test.to_csv('binary_x_test.csv', index =0)
y_train.to_csv('binary_y_train.csv', index = 0)
y_test.to_csv('binary_y_test.csv', index = 0)


train_df = pd.concat([x_train.reset_index(drop=True), y_train], axis=1)
test_df  = pd.concat([x_test.reset_index(drop=True), y_test], axis=1)


train_df.to_csv('binary_train.csv', index =0)
test_df.to_csv('binary_test.csv', index = 0)
